style(YALB-72): defined button focus styles
